fix bug auth
