+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "free-covid-consultation-calendar",
-  "description": "Example Gatsby, and Netlify CMS project",
+  "description": "Free Video Health Consultations",
   "version": "1.1.3",
-  "author": "Austin Green",
+  "author": "Marcos Cannabrava, Stu Wood",
   "dependencies": {
     "bulma": "^0.8.0",
     "core-js": "^2.6.5",
