+++ src/components/LayoutTemplate.js
@@ -1,0 +1,53 @@
+import React from 'react'
+import { Helmet } from 'react-helmet'
+import './all.sass'
+import useSiteMetadata from './SiteMetadata'
+import { withPrefix } from 'gatsby'
+
+const TemplateWrapper = () => {
+  const { title, description } = useSiteMetadata()
+  return (
+    <div>
+      <Helmet>
+        <html lang="en" />
+        <title>{title}</title>
+        <meta name="description" content={description} />
+
+        <link
+          rel="apple-touch-icon"
+          sizes="180x180"
+          href={`${withPrefix('/')}img/apple-touch-icon.png`}
+        />
+        <link
+          rel="icon"
+          type="image/png"
+          href={`${withPrefix('/')}img/stethoscope-icon.png`}//32x32 is preferred
+          sizes="32x32"
+        />
+        <link
+          rel="icon"
+          type="image/png"
+          href={`${withPrefix('/')}img/stethoscope-icon.png`}//16x16 is preferred
+          sizes="16x16"
+        />
+
+        <link
+          rel="mask-icon"
+          href={`${withPrefix('/')}img/safari-pinned-tab.svg`}
+          color="#ff4400"
+        />
+        <meta name="theme-color" content="#fff" />
+
+        <meta property="og:type" content="business.business" />
+        <meta property="og:title" content={title} />
+        <meta property="og:url" content="/" />
+        <meta
+          property="og:image"
+          content={`${withPrefix('/')}img/og-image.jpg`}
+        />
+      </Helmet>
+    </div>
+  )
+}
+
+export default TemplateWrapper
