+++ src/pages/account.js
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 import { compose } from 'recompose';
 
-import Layout from '../components/layout';
+import Layout from '../components/Layout';
 import {
   AuthUserContext,
   withAuthorization,
