+++ src/pages/signup.js
@@ -1,6 +1,6 @@
 import React, { Fragment } from 'react';
 
-import Layout from '../components/layout';
+import Layout from '../components/Layout';
 import SignUpForm from '../components/SignUp';
 
 const SignUpPage = () => (
